Moved RobotWidgetTree class to dedicate file.
