+++ src/yarpmotorgui/modestreemanager.h
@@ -16,8 +16,7 @@
 #include <string>
 #include <unordered_map>
 #include "partitem.h"
-#include "partItemTree.h"
-#include "customgroupbox.h"
+#include "robotWidgetTree.h"
 
 class ModesListWidget : public QTreeWidget
 {
@@ -25,31 +24,6 @@ class ModesListWidget : public QTreeWidget
 
 public:
     ModesListWidget(QWidget * parent = 0);
-};
-
-class RobotWidgetTree : public CustomGroupBox
-{
-    Q_OBJECT
-
-    struct PartPointers
-    {
-        PartItemTree* partWidget;
-        CustomGroupBox* partWidgetParent;
-    };
-
-    std::unordered_map<int, PartPointers> m_indexToPartMap;
-    QIcon m_okIcon;
-    QIcon m_warningIcon;
-
-public:
-
-    void setIcons(const QIcon& okIcon, const QIcon& warningIcon);
-
-    void addPart(const std::string &partName, int partIndex, PartItemTree* partTreeWidget);
-
-    void updateRobotPart(int index, const QVector<JointItem::JointState>& modes);
-
-    void resizeEvent(QResizeEvent *event) override;
 };
 
 class ModesTreeManager : public QObject
